feat(bridge): Adapt to PR walk-through change requests
