+++ bridge/client/app/_components/ktb-expandable-tile/ktb-expandable-tile.module.ts
@@ -1,6 +1,5 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { DtExpandablePanelModule } from '@dynatrace/barista-components/expandable-panel';
 import { DtIconModule } from '@dynatrace/barista-components/icon';
 import { DtShowMoreModule } from '@dynatrace/barista-components/show-more';
@@ -10,7 +9,6 @@ import { KtbExpandableTileComponent, KtbExpandableTileHeaderDirective } from './
   declarations: [KtbExpandableTileComponent, KtbExpandableTileHeaderDirective],
   imports: [
     CommonModule,
-    BrowserAnimationsModule,
     DtExpandablePanelModule,
     DtIconModule.forRoot({
       svgIconLocation: `assets/icons/{{name}}.svg`,
