+++ bridge/client/app/_components/ktb-heatmap/ktb-heatmap.module.ts
@@ -1,15 +1,15 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { KtbHeatmapComponent } from './ktb-heatmap.component';
-import { KtbHeatmapTooltipComponent } from './ktb-heatmap-tooltip.component';
+import { HttpClientModule } from '@angular/common/http';
+import { NgModule } from '@angular/core';
+import { DtButtonModule } from '@dynatrace/barista-components/button';
+import { DtIconModule } from '@dynatrace/barista-components/icon';
 import { DtKeyValueListModule } from '@dynatrace/barista-components/key-value-list';
 import { KtbPipeModule } from '../../_pipes/ktb-pipe.module';
-import { DtIconModule } from '@dynatrace/barista-components/icon';
-import { DtButtonModule } from '@dynatrace/barista-components/button';
-import { HttpClientModule } from '@angular/common/http';
+import { KtbHeatmapTooltipComponent } from './ktb-heatmap-tooltip.component';
+import { KtbHeatmapComponent } from './ktb-heatmap.component';
 
 @NgModule({
-  declarations: [KtbHeatmapComponent, KtbHeatmapTooltipComponent], // add KtbTestHeatmapComponent for testing
+  declarations: [KtbHeatmapComponent, KtbHeatmapTooltipComponent],
   imports: [
     CommonModule,
     HttpClientModule,
@@ -20,6 +20,6 @@ import { HttpClientModule } from '@angular/common/http';
     }),
     DtButtonModule,
   ],
-  exports: [KtbHeatmapComponent, KtbHeatmapTooltipComponent], // add KtbTestHeatmapComponent for testing
+  exports: [KtbHeatmapComponent, KtbHeatmapTooltipComponent],
 })
 export class KtbHeatmapModule {}
