--- bridge/client/app/_components/ktb-heatmap/testing/ktb-test-heatmap.component.html
@@ -1,4 +1,0 @@
-<div class="heatmap-container">
-  <ktb-heatmap [dataPoints]="dataPoints" [selectedIdentifier]="selectedIdentifier"> </ktb-heatmap>
-  <ktb-heatmap [dataPoints]="dataPoints" class="mt-3"> </ktb-heatmap>
-</div>
