--- bridge/client/app/_components/ktb-heatmap/testing/ktb-test-heatmap.component.scss
@@ -1,4 +1,0 @@
-.heatmap-container {
-  margin: 50px 50px;
-  width: 80%;
-}
