+++ bridge/client/app/_components/ktb-webhook-settings/ktb-webhook-settings.module.ts
@@ -1,5 +1,4 @@
 import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { ReactiveFormsModule } from '@angular/forms';
@@ -16,7 +15,6 @@ import { KtbWebhookSettingsComponent } from './ktb-webhook-settings.component';
   declarations: [KtbWebhookSettingsComponent],
   imports: [
     CommonModule,
-    HttpClientModule,
     DtButtonModule,
     DtSelectModule,
     DtIconModule.forRoot({
