+++ bridge/package.json
@@ -59,7 +59,7 @@
     "@dynatrace/barista-components": "10.7.5",
     "@dynatrace/barista-fonts": "1.0.1",
     "@dynatrace/barista-icons": "7.23.0",
-    "@storybook/angular": "^6.5.9",
+    "@storybook/angular": "6.5.9",
     "cypress-file-upload": "5.0.8",
     "d3": "7.6.1",
     "d3-scale": "4.0.2",
@@ -85,16 +85,16 @@
     "@angular/cli": "12.2.18",
     "@angular/compiler-cli": "12.2.16",
     "@angular/language-service": "12.2.16",
-    "@babel/core": "^7.18.6",
-    "@compodoc/compodoc": "^1.1.11",
+    "@babel/core": "7.18.6",
+    "@compodoc/compodoc": "1.1.11",
     "@cypress/schematic": "2.0.0",
-    "@storybook/addon-actions": "^6.5.9",
-    "@storybook/addon-essentials": "^6.5.9",
-    "@storybook/addon-interactions": "^6.5.9",
-    "@storybook/addon-links": "^6.5.9",
-    "@storybook/builder-webpack5": "^6.5.9",
-    "@storybook/manager-webpack5": "^6.5.9",
-    "@storybook/testing-library": "^0.0.13",
+    "@storybook/addon-actions": "6.5.9",
+    "@storybook/addon-essentials": "6.5.9",
+    "@storybook/addon-interactions": "6.5.9",
+    "@storybook/addon-links": "6.5.9",
+    "@storybook/builder-webpack5": "6.5.9",
+    "@storybook/manager-webpack5": "6.5.9",
+    "@storybook/testing-library": "0.0.13",
     "@types/d3": "7.4.0",
     "@types/dompurify": "2.3.3",
     "@types/jest": "28.1.6",
@@ -106,7 +106,7 @@
     "@types/uuid": "8.3.4",
     "@typescript-eslint/eslint-plugin": "5.27.1",
     "@typescript-eslint/parser": "5.27.1",
-    "babel-loader": "^8.2.5",
+    "babel-loader": "8.2.5",
     "cypress": "10.3.0",
     "eslint": "7.32.0",
     "eslint-config-airbnb-base": "15.0.0",
