+++ bridge/stories/components/Components.stories.mdx
@@ -42,3 +42,7 @@ with different inputs easily this should be the playground for component integra
 ## Danger Zone
 
 <Story id="components-danger-zone--project" />
+
+<div>
+  <h1>Blubb</h1>
+</div>
